Add a fix to dropdown menus for the commmunities UI
Blizzard really did a nasty number here, with not initializing even a single secure menu while opening the frame. Most of the time they'll initialize something, at least.

It's still theoretically possible to run into sequences of events that trigger an update to those menus without an initialization happening, but this prevents it from happening by default on the pane opening.